libap: implement internal file-locking logic
This implements file-locking logic compatible with liblockfile
`lockfile_create()` and `lockfile_remove()`.
It is a simplified port of those, tailored to libap needs.
The goal of this change is to allow removing the direct
dependency on `liblockfile`, which is a library mainly targeted
at mailbox locking and comes with a set-gid binary.
Neither of those are actually useful for s390-tools usage, and
cause further troubles to downstream distributions.